Updates for table changes, began work on staff_worker functions
Updated table to get rid of the join linking "advises", "advisee", and
"advisor" in favor of just have an "advisor_id" attribute on advisee.
Also started putting together some of the functions for staff worker and
updated side bar a bit.  Also added database filling functions to
dashboard as well as advisor for use with database performance testing.